rate table tests completed
