+++ Src/DfT.DTRO/Enums/CurrencyType.cs
@@ -1,0 +1,19 @@
+﻿namespace DfT.DTRO.Enums;
+
+/// <summary>
+/// Currency type
+/// </summary>
+public enum CurrencyType
+{
+    /// <summary>
+    /// Euro currency type
+    /// </summary>
+    [Display(Name = "EUR")]
+    Eur,
+
+    /// <summary>
+    /// Great Britain pound type
+    /// </summary>
+    [Display(Name = "GBP")]
+    Gbp
+}
