+++ Src/DfT.DTRO/Enums/RateType.cs
@@ -1,0 +1,19 @@
+﻿namespace DfT.DTRO.Enums;
+
+/// <summary>
+/// Rate type
+/// </summary>
+public enum RateType
+{
+    /// <summary>
+    /// Daily rate type
+    /// </summary>
+    [Display(Name = "daily")]
+    Daily,
+
+    /// <summary>
+    /// Hourly rate type
+    /// </summary>
+    [Display(Name = "hourly")]
+    Hourly
+}
