+++ Src/DfT.DTRO/Services/Validation/Contracts/IConditionValidationService.cs
@@ -1,9 +1,15 @@
-﻿using DfT.DTRO.Models.Conditions;
-using DfT.DTRO.Models.Validation;
+﻿namespace DfT.DTRO.Services.Validation.Contracts;
 
-namespace DfT.DTRO.Services.Validation.Contracts;
-
+/// <summary>
+/// Validation Condition Service
+/// </summary>
 public interface IConditionValidationService
 {
-    public List<SemanticValidationError> Validate(ConditionSet conditions);
+    /// <summary>
+    /// Check conditions
+    /// </summary>
+    /// <param name="dtroSubmit">D-TRO payload to check.</param>
+    /// <param name="schemaVersion">Schema version passed in.</param>
+    /// <returns>List of validation errors</returns>
+    List<SemanticValidationError> ValidateCondition(DtroSubmit dtroSubmit, SchemaVersion schemaVersion);
 }
