+++ Src/DfT.DTRO/Services/Validation/Contracts/IOldConditionValidationService.cs
@@ -1,0 +1,8 @@
+﻿using DfT.DTRO.Models.Conditions;
+
+namespace DfT.DTRO.Services.Validation.Contracts;
+
+public interface IOldConditionValidationService
+{
+    public List<SemanticValidationError> Validate(ConditionSet conditions);
+}
